requirements.txt
streamlit
pandas
python-dotenv
google-generative-ai
pyarrow
numpy
